reorganize kernels structure
