+++ mgplvm/kernels/__init__.py
@@ -1,0 +1,3 @@
+from .kernel import Kernel
+from .stationary import (QuadExp, Exp, Matern)
+from .linear import Linear
